patch 7.4.845 Problem:    Compiler warning for possible loss of data. Solution:   Add a type cast. (Erich Ritz)

+++ src/misc1.c
@@ -3985,7 +3985,7 @@ expand_env_esc(srcp, dst, dstlen, esc, one, startstr)
 		len = dstlen;
 	    vim_strncpy(dst, var, len);
 	    dst += len;
-	    dstlen -= len;
+	    dstlen -= (int)len;
 	    continue;
 	}
 #endif
